Trekker ut "reduce"-funksjonaliteten til en egen funksjon og legger den i ny fil  utils.ts
